resultats/repositories/equipe_10_ivoice/ivoice-main/requirements.txt
python-dotenv==1.0.1
langchain==0.3.7
langchain-core==0.3.19
langchain-text-splitters==0.3.2
langchain-community==0.3.7
openai==1.55.0
dacite==1.8.1
streamlit==1.40.1
